requirements.txt
google-cloud-speech
google-cloud-translate
